% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{qfmt}
\alias{qfmt}
\title{Format a list to use as http query}
\usage{
qfmt(x)
}
\arguments{
\item{x}{list, where values are to be formatted}
}
\value{
list
}
\description{
Format a list to use as http query
}

