% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_atomic.R
\docType{methods}
\name{serialize_atomic}
\alias{serialize_atomic}
\alias{serialize_atomic,Duration-method}
\alias{serialize_atomic,Period-method}
\alias{serialize_atomic.default}
\alias{serialize_atomic.difftime}
\alias{serialize_atomic.logical}
\title{Serialize an atomic vector}
\usage{
serialize_atomic(x, locale = default_locale(), ...)

serialize_atomic.default(x, locale = default_locale(), ...)

serialize_atomic.logical(x, locale = default_locale(), ...)

serialize_atomic.difftime(x, locale = default_locale(), ...)

\S4method{serialize_atomic}{Duration}(x, locale = default_locale(), ...)

\S4method{serialize_atomic}{Period}(x, locale = default_locale(), ...)
}
\arguments{
\item{x}{atomic vector}

\item{locale}{list, generated using \code{\link{locale}}}

\item{...}{additional arguments}
}
\value{
character
}
\description{
Serialize an atomic vector
}
\keyword{internal}

