% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_list.R
\name{serialize_list}
\alias{serialize_list}
\title{Serialize a list into text}
\usage{
serialize_list(x, delim = ",", locale = default_locale())
}
\arguments{
\item{x}{list, values to be serialized}

\item{delim}{character, delimiter to collapse mutiple-value members}

\item{locale}{list, directs the formatting functions for each class}
}
\value{
list with same names as \code{x}, all values are \code{\link{character}}
}
\description{
This function acts as a text-serializer for lists. One use case is
}
\examples{
library("lubridate")

query_param <- list(
  delay = dseconds(4),
  next_steps = c("collate", "send"),
  verbose = TRUE
)

serialize_list(query_param)

serialize_list(query_param, delim = " ")
}
\seealso{
\code{\link{locale}}
}

