% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{srlst}
\alias{srlst}
\title{Serialize a list into text}
\usage{
srlst(x, delim = ",", locale = default_locale())
}
\arguments{
\item{x}{list, where values are to be formatted (serialized)}

\item{delim}{character, delimiter to collapse mutiple-value members}

\item{locale}{list, directs the formatting functions for each class}
}
\value{
list with same names as \code{x}, values are all \code{\link{character}}
}
\description{
This function acts as a text-serializer for lists. One use case is
}

