% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_atomic.R
\docType{methods}
\name{srlz_atomic}
\alias{srlz_atomic}
\alias{srlz_atomic,Duration-method}
\alias{srlz_atomic,Period-method}
\alias{srlz_atomic.default}
\alias{srlz_atomic.difftime}
\alias{srlz_atomic.logical}
\title{Format an atomic vector}
\usage{
srlz_atomic(x, locale = default_locale(), ...)

srlz_atomic.default(x, locale = default_locale(), ...)

srlz_atomic.logical(x, locale = default_locale(), ...)

srlz_atomic(x, locale = default_locale(), ...)

\S4method{srlz_atomic}{Duration}(x, locale = default_locale(), ...)

\S4method{srlz_atomic}{Period}(x, locale = default_locale(), ...)

srlz_atomic.difftime(x, locale = default_locale(), ...)
}
\arguments{
\item{x}{atomic vector}

\item{locale}{list, generated using \code{\link{locale}}}

\item{...}{additional arguments}
}
\value{
character
}
\description{
Format an atomic vector
}
\keyword{internal}

