% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{locale}
\alias{locale}
\title{Create a locale}
\usage{
locale(logical_format = list(true = "true", false = "false"),
  dtime_format = c("ms", "s", "minute", "hour", "day", "week"))
}
\arguments{
\item{logical_format}{list, with two members: \code{true} and \code{false},
character strings indicating how logicals are written.}

\item{dtime_format}{character, indicating how to express time-differences.
Using \code{\link{srlst}}, time-differences are indicated by using
\code{\link{difftime}}, \code{lubridate::\link[lubridate]{duration}}, or
\code{lubridate::\link[lubridate]{period}}.
The options indicate which time unit to use, \code{"ms"} being the default.
The output for time-differences is a character representation of the numeric
value of the time-differences expressed using that unit.}
}
\description{
This locale object acts much like the locale object from the
\strong{readr} package. It is used to provide direction to the
formatting methods for classes of each individual memeber of the
query-parameter list sent to \code{\link{srlst}}.
}
\seealso{
\code{\link{srlst}}
}

