% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{locale}
\alias{locale}
\title{Create a locale}
\usage{
locale(logical_format = list(true = "true", false = "false"),
  dtime_format = c("ms", "s", "minute", "hour", "day", "week"))
}
\arguments{
\item{logical_format}{list, with two members: \code{true} and \code{false},
indicating how logicals will be written.}

\item{dtime_format}{character, indicates how to express time-differences.}
}
\value{
list
}
\description{
This locale object acts much like the locale object from the \strong{readr} package.
It is used to provide direction to the formatting methods for individual classes.
}

