% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{qfmt}
\alias{qfmt}
\title{Format a list to use as http query}
\usage{
qfmt(x, delim = ",", locale = default_locale())
}
\arguments{
\item{x}{list, where values are to be formatted}

\item{delim}{character, delimiter for mutiple-value members}

\item{locale}{list, directs the formatting functions for each class}
}
\value{
list
}
\description{
Format a list to use as http query
}

