% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfmtr_fmt.R
\docType{methods}
\name{qfmtr_fmt}
\alias{qfmtr_fmt}
\alias{qfmtr_fmt,Duration-method}
\alias{qfmtr_fmt,Period-method}
\alias{qfmtr_fmt.default}
\alias{qfmtr_fmt.logical}
\title{Format an atomic vector}
\usage{
qfmtr_fmt(x, ...)

qfmtr_fmt.default(x, ...)

qfmtr_fmt.logical(x, ...)

qfmtr_fmt(x, ...)

\S4method{qfmtr_fmt}{Duration}(x, ...)

\S4method{qfmtr_fmt}{Period}(x, ...)
}
\arguments{
\item{x}{atomic vector}

\item{...}{additional arguments}
}
\value{
character
}
\description{
Format an atomic vector
}
\keyword{internal}

